requirements.txt
# Core ML/DL
tensorflow>=2.15.0
keras>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0

# Web application
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pandas>=2.0.0
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
